requirements.txt
flask
requests
msgspec
python-dotenv